requirements.txt
asgiref==3.7.2
Django==4.2.7
django-crontab==0.7.1
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4